+++ docs/changelog.md
@@ -1,12 +1,35 @@
 # TBD
 
+### Features
+- The "Starlark code successfully executed" or "Error encountered running Starlark code" messages are now "Starlark 
+code successfully run in dry-run mode" and "Error encountered running Starlark code in dry-run mode" when Starlark is 
+run in dry-run mode (and without the "in dry-run mode" when the script is executed for real)
+
+### Fixes
+- Don't duplicate instruction position information in `store_service_files`
+
+### Removals
+- Remove stack trace from validation errors as it isn't used currently
+
 ### Changes
-- Replaced "insternal ID" in the output message of `add_service` and `remove_service` instructions with "service GUID"
+- Changed validation message from "Pre-validating" to "Validating"
+- Disabled progress info in non-interactive terminals when running a Starlark Package
+
+# 0.57.5
+### Changes
+- Replaced stack name with the stack file name in custom evaluation errors
+- Replaced "internal ID" in the output message of `add_service` and `remove_service` instructions with "service GUID"
+
+### Features
+- Support public ports in Starlark to cover the NEAR usecase
 
 ### Fixes
 - Corrected some old references to Starlark "modules"
 - Fixed a typo where the CLI setup URL was redirecting to the CI setup
 - Corrected almost all old references to `docs.kurtosistech.com`
+- Changed the name from startosis to starlark in the `internal_testsuite` build script
+- Fixed `internal-testsuites` omission during build time
+- Fixed a bug related to omitting the `enclave ID` value when a function which filters modules is called
 
 # 0.57.4
 ### Changes